add calcite adder

+++ lyb-other-1.12/src/main/java/hbase/HBaseAsyncUtil.java
@@ -1,5 +1,9 @@
 package hbase;
 
+import java.io.IOException;
+import java.io.Serializable;
+import java.text.DecimalFormat;
+import java.util.*;
 import org.apache.commons.lang.StringUtils;
 import org.apache.hadoop.conf.Configuration;
 import org.apache.hadoop.hbase.*;
@@ -12,15 +16,7 @@ import org.apache.hadoop.hbase.util.Bytes;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import java.io.IOException;
-import java.io.Serializable;
-import java.text.DecimalFormat;
-import java.util.*;
-
-/**
- * HBase Util
- *  2018/07/26
- */
+/** HBase Util 2018/07/26 */
 public class HBaseAsyncUtil {
 
     private static final Logger logger = LoggerFactory.getLogger(HBaseAsyncUtil.class);
@@ -32,7 +28,8 @@ public class HBaseAsyncUtil {
         try {
             if (conf == null) {
                 conf = HBaseConfiguration.create();
-//        conf.set("hbase.zookeeper.property.clientPort", ConfigUtil.getInstance().getConfigVal("zkport", ConstantProperties.COMMON_PROP));
+                //        conf.set("hbase.zookeeper.property.clientPort",
+                // ConfigUtil.getInstance().getConfigVal("zkport", ConstantProperties.COMMON_PROP));
                 conf.set("hbase.zookeeper.quorum", zkHost);
                 conf.set("zookeeper.znode.parent", "/hbase");
             }
@@ -42,20 +39,17 @@ public class HBaseAsyncUtil {
         }
     }
 
-    /**
-     * 获取连接
-     */
+    /** 获取连接 */
     public static synchronized Connection getConnection() {
         try {
             if (conn == null || conn.isClosed()) {
                 conn = ConnectionFactory.createConnection(conf);
             }
-//     System.out.println("---------- " + conn.hashCode());
+            //     System.out.println("---------- " + conn.hashCode());
         } catch (IOException e) {
             logger.error("HBase 建立连接失败 ", e);
         }
         return conn;
-
     }
 
     /**
@@ -64,9 +58,14 @@ public class HBaseAsyncUtil {
      * @param tableName
      * @throws Exception
      */
-    public static void createTable(String tableName, String[] columnFamilies, boolean preBuildRegion) throws Exception {
+    public static void createTable(
+            String tableName, String[] columnFamilies, boolean preBuildRegion) throws Exception {
         if (preBuildRegion) {
-            String[] s = new String[]{"sta.txt", "2", "3", "4", "5", "6", "7", "8", "9", "A", "B", "C", "D", "E", "F"};
+            String[] s =
+                    new String[] {
+                        "sta.txt", "2", "3", "4", "5", "6", "7", "8", "9", "A", "B", "C", "D", "E",
+                        "F"
+                    };
             int partition = 16;
             byte[][] splitKeys = new byte[partition - 1][];
             for (int i = 1; i < partition; i++) {
@@ -84,7 +83,7 @@ public class HBaseAsyncUtil {
         for (int i = 1; i <= pNum; i++) {
             splitKeys[i - 1] = Bytes.toBytes(s[i - 1]);
         }
-        createTable(tableName, new String[]{"events"}, splitKeys);
+        createTable(tableName, new String[] {"events"}, splitKeys);
     }
 
     /**
@@ -94,7 +93,8 @@ public class HBaseAsyncUtil {
      * @param cfs
      * @throws IOException
      */
-    private static void createTable(String tableName, String[] cfs, byte[][] splitKeys) throws Exception {
+    private static void createTable(String tableName, String[] cfs, byte[][] splitKeys)
+            throws Exception {
         Connection conn = getConnection();
         HBaseAdmin admin = (HBaseAdmin) conn.getAdmin();
         try {
@@ -190,7 +190,7 @@ public class HBaseAsyncUtil {
      * 给 table 创建 snapshot
      *
      * @param snapshotName 快照名称
-     * @param tableName    表名
+     * @param tableName 表名
      * @return
      * @throws IOException
      */
@@ -215,8 +215,7 @@ public class HBaseAsyncUtil {
             Admin admin = getConnection().getAdmin();
             if (StringUtils.isNotBlank(snapshotNameRegex))
                 return admin.listSnapshots(snapshotNameRegex);
-            else
-                return admin.listSnapshots();
+            else return admin.listSnapshots();
         } catch (Exception e) {
             logger.error("Snapshot " + snapshotNameRegex + " get failed !", e);
         }
@@ -233,10 +232,8 @@ public class HBaseAsyncUtil {
     public static void deleteSnapshots(String snapshotNameRegex) {
         try {
             Admin admin = getConnection().getAdmin();
-            if (StringUtils.isNotBlank(snapshotNameRegex))
-                admin.deleteSnapshots(snapshotNameRegex);
-            else
-                logger.error("SnapshotNameRegex can't be null !");
+            if (StringUtils.isNotBlank(snapshotNameRegex)) admin.deleteSnapshots(snapshotNameRegex);
+            else logger.error("SnapshotNameRegex can't be null !");
         } catch (Exception e) {
             logger.error("Snapshots " + snapshotNameRegex + " del failed !", e);
         }
@@ -252,10 +249,8 @@ public class HBaseAsyncUtil {
     public static void deleteSnapshot(String snapshotName) {
         try {
             Admin admin = getConnection().getAdmin();
-            if (StringUtils.isNotBlank(snapshotName))
-                admin.deleteSnapshot(snapshotName);
-            else
-                logger.error("SnapshotName can't be null !");
+            if (StringUtils.isNotBlank(snapshotName)) admin.deleteSnapshot(snapshotName);
+            else logger.error("SnapshotName can't be null !");
         } catch (Exception e) {
             logger.error("Snapshot " + snapshotName + " del failed !", e);
         }
@@ -265,20 +260,26 @@ public class HBaseAsyncUtil {
      * 分页检索表数据。<br>
      * （如果在创建表时为此表指定了非默认的命名空间，则需拼写上命名空间名称，格式为【namespace:tablename】）。
      *
-     * @param tableName   表名称(*)。
+     * @param tableName 表名称(*)。
      * @param startRowKey 起始行键(可以为空，如果为空，则从表中第一行开始检索)。
-     * @param endRowKey   结束行键(可以为空)。
-     * @param filterList  检索条件过滤器集合(不包含分页过滤器；可以为空)。
+     * @param endRowKey 结束行键(可以为空)。
+     * @param filterList 检索条件过滤器集合(不包含分页过滤器；可以为空)。
      * @param maxVersions 指定最大版本数【如果为最大整数值，则检索所有版本；如果为最小整数值，则检索最新版本；否则只检索指定的版本数】。
-     * @param pageModel   分页模型(*)。
+     * @param pageModel 分页模型(*)。
      * @return 返回HBasePageModel分页对象。
      */
-    public static HBasePageModel scanResultByPageFilter(String tableName, byte[] startRowKey, byte[] endRowKey, FilterList filterList, int maxVersions, HBasePageModel pageModel) {
+    public static HBasePageModel scanResultByPageFilter(
+            String tableName,
+            byte[] startRowKey,
+            byte[] endRowKey,
+            FilterList filterList,
+            int maxVersions,
+            HBasePageModel pageModel) {
         if (pageModel == null) {
             pageModel = new HBasePageModel(10);
         }
         if (maxVersions <= 0) {
-//默认只检索数据的最新版本
+            // 默认只检索数据的最新版本
             maxVersions = Integer.MIN_VALUE;
         }
         pageModel.initStartTime();
@@ -293,7 +294,7 @@ public class HBaseAsyncUtil {
             int tempPageSize = pageModel.getPageSize();
             boolean isEmptyStartRowKey = false;
             if (startRowKey == null) {
-//则读取表的第一行记录
+                // 则读取表的第一行记录
                 Result firstResult = selectFirstResultRow(tableName, filterList);
                 if (firstResult.isEmpty()) {
                     return pageModel;
@@ -307,7 +308,7 @@ public class HBaseAsyncUtil {
                 if (pageModel.getPageEndRowKey() != null) {
                     pageModel.setPageStartRowKey(pageModel.getPageEndRowKey());
                 }
-//从第二页开始，每次都多取一条记录，因为第一条记录是要删除的。
+                // 从第二页开始，每次都多取一条记录，因为第一条记录是要删除的。
                 tempPageSize += 1;
             }
 
@@ -358,9 +359,10 @@ public class HBaseAsyncUtil {
         int pageIndex = pageModel.getPageIndex() + 1;
         pageModel.setPageIndex(pageIndex);
         if (pageModel.getResultList().size() > 0) {
-//获取本次分页数据首行和末行的行键信息
+            // 获取本次分页数据首行和末行的行键信息
             byte[] pageStartRowKey = pageModel.getResultList().get(0).getRow();
-            byte[] pageEndRowKey = pageModel.getResultList().get(pageModel.getResultList().size() - 1).getRow();
+            byte[] pageEndRowKey =
+                    pageModel.getResultList().get(pageModel.getResultList().size() - 1).getRow();
             pageModel.setPageStartRowKey(pageStartRowKey);
             pageModel.setPageEndRowKey(pageEndRowKey);
         }
@@ -375,7 +377,7 @@ public class HBaseAsyncUtil {
      * 检索指定表的第一行记录。<br>
      * （如果在创建表时为此表指定了非默认的命名空间，则需拼写上命名空间名称，格式为【namespace:tablename】）。
      *
-     * @param tableName  表名称(*)。
+     * @param tableName 表名称(*)。
      * @param filterList 过滤器集合，可以为null。
      * @return
      */
@@ -414,24 +416,26 @@ public class HBaseAsyncUtil {
      * 异步往指定表添加数据
      *
      * @param tablename 表名
-     * @param puts      需要添加的数据
-     * @return long     返回执行时间
+     * @param puts 需要添加的数据
+     * @return long 返回执行时间
      * @throws IOException
      */
     public static long put(String tablename, List<SocPut> puts) throws Exception {
         long currentTime = System.currentTimeMillis();
         Connection conn = getConnection();
-        final BufferedMutator.ExceptionListener listener = new BufferedMutator.ExceptionListener() {
-            @Override
-            public void onException(RetriesExhaustedWithDetailsException e, BufferedMutator mutator) {
-                for (int i = 0; i < e.getNumExceptions(); i++) {
-                    System.out.println("Failed to sent put " + e.getRow(i) + ".");
-                    logger.error("Failed to sent put " + e.getRow(i) + ".");
-                }
-            }
-        };
-        BufferedMutatorParams params = new BufferedMutatorParams(TableName.valueOf(tablename))
-                .listener(listener);
+        final BufferedMutator.ExceptionListener listener =
+                new BufferedMutator.ExceptionListener() {
+                    @Override
+                    public void onException(
+                            RetriesExhaustedWithDetailsException e, BufferedMutator mutator) {
+                        for (int i = 0; i < e.getNumExceptions(); i++) {
+                            System.out.println("Failed to sent put " + e.getRow(i) + ".");
+                            logger.error("Failed to sent put " + e.getRow(i) + ".");
+                        }
+                    }
+                };
+        BufferedMutatorParams params =
+                new BufferedMutatorParams(TableName.valueOf(tablename)).listener(listener);
         params.writeBufferSize(5 * 1024 * 1024);
 
         final BufferedMutator mutator = conn.getBufferedMutator(params);
@@ -449,8 +453,8 @@ public class HBaseAsyncUtil {
      * 异步往指定表添加数据
      *
      * @param tablename 表名
-     * @param put       需要添加的数据
-     * @return long     返回执行时间
+     * @param put 需要添加的数据
+     * @return long 返回执行时间
      * @throws IOException
      */
     public static long put(String tablename, SocPut put) throws Exception {
@@ -461,8 +465,8 @@ public class HBaseAsyncUtil {
      * 往指定表添加数据
      *
      * @param tablename 表名
-     * @param puts      需要添加的数据
-     * @return long     返回执行时间
+     * @param puts 需要添加的数据
+     * @return long 返回执行时间
      * @throws IOException
      */
     public static long putByHTable(String tablename, List<?> puts) throws Exception {
@@ -533,7 +537,7 @@ public class HBaseAsyncUtil {
     public static void closeConnect(Connection conn) {
         if (null != conn) {
             try {
-//     conn.close();
+                //     conn.close();
             } catch (Exception e) {
                 logger.error("closeConnect failure !", e);
             }
@@ -634,9 +638,7 @@ public class HBaseAsyncUtil {
         return results;
     }
 
-    /**
-     * 格式化输出结果
-     */
+    /** 格式化输出结果 */
     public static void formatRow(KeyValue[] rs) {
         for (KeyValue kv : rs) {
             System.out.println(" column family : " + Bytes.toString(kv.getFamily()));
@@ -661,14 +663,13 @@ public class HBaseAsyncUtil {
         }
         return num;
     }
-
 }
-
 
 /**
  * Description: HBase表数据分页模型类。<br>
  * 利用此类可管理多个HBaseQualifierModel对象。
- * @author 
+ *
+ * @author
  * @version 1.0
  */
 class HBasePageModel implements Serializable {
@@ -689,11 +690,13 @@ class HBasePageModel implements Serializable {
     private long startTime = System.currentTimeMillis();
     private long endTime = System.currentTimeMillis();
     private List<Result> resultList = new ArrayList<Result>();
+
     public HBasePageModel(int pageSize) {
         this.pageSize = pageSize;
     }
     /**
      * 获取分页记录数量
+     *
      * @return
      */
     public int getPageSize() {
@@ -701,6 +704,7 @@ class HBasePageModel implements Serializable {
     }
     /**
      * 设置分页记录数量
+     *
      * @param pageSize
      */
     public void setPageSize(int pageSize) {
@@ -708,6 +712,7 @@ class HBasePageModel implements Serializable {
     }
     /**
      * 获取当前页序号
+     *
      * @return
      */
     public int getPageIndex() {
@@ -715,6 +720,7 @@ class HBasePageModel implements Serializable {
     }
     /**
      * 设置当前页序号
+     *
      * @param pageIndex
      */
     public void setPageIndex(int pageIndex) {
@@ -722,6 +728,7 @@ class HBasePageModel implements Serializable {
     }
     /**
      * 获取分页总数
+     *
      * @return
      */
     public int getPageCount() {
@@ -729,6 +736,7 @@ class HBasePageModel implements Serializable {
     }
     /**
      * 设置分页总数
+     *
      * @param pageCount
      */
     public void setPageCount(int pageCount) {
@@ -736,6 +744,7 @@ class HBasePageModel implements Serializable {
     }
     /**
      * 获取每页的第一行序号
+     *
      * @return
      */
     public int getPageFirstRowIndex() {
@@ -744,6 +753,7 @@ class HBasePageModel implements Serializable {
     }
     /**
      * 获取每页起始行键
+     *
      * @return
      */
     public byte[] getPageStartRowKey() {
@@ -751,6 +761,7 @@ class HBasePageModel implements Serializable {
     }
     /**
      * 设置每页起始行键
+     *
      * @param pageStartRowKey
      */
     public void setPageStartRowKey(byte[] pageStartRowKey) {
@@ -758,23 +769,23 @@ class HBasePageModel implements Serializable {
     }
     /**
      * 获取每页结束行键
+     *
      * @return
      */
     public byte[] getPageEndRowKey() {
         return pageEndRowKey;
     }
-    /**
-     * 设置每页结束行键
-     */
+    /** 设置每页结束行键 */
     public void setPageEndRowKey(byte[] pageEndRowKey) {
         this.pageEndRowKey = pageEndRowKey;
     }
     /**
      * 获取上一页序号
+     *
      * @return
      */
     public int getPrevPageIndex() {
-        if(this.getPageIndex() > 1) {
+        if (this.getPageIndex() > 1) {
             this.prevPageIndex = this.getPageIndex() - 1;
         } else {
             this.prevPageIndex = 1;
@@ -783,6 +794,7 @@ class HBasePageModel implements Serializable {
     }
     /**
      * 获取下一页序号
+     *
      * @return
      */
     public int getNextPageIndex() {
@@ -791,44 +803,41 @@ class HBasePageModel implements Serializable {
     }
     /**
      * 获取是否有下一页
+     *
      * @return
      */
     public boolean isHasNextPage() {
-//这个判断是不严谨的，因为很有可能剩余的数据刚好够一页。
-        if(this.getResultList().size() == this.getPageSize()) {
+        // 这个判断是不严谨的，因为很有可能剩余的数据刚好够一页。
+        if (this.getResultList().size() == this.getPageSize()) {
             this.hasNextPage = true;
         } else {
             this.hasNextPage = false;
         }
         return hasNextPage;
     }
-    /**
-     * 获取已检索总记录数
-     */
+    /** 获取已检索总记录数 */
     public int getQueryTotalCount() {
         return queryTotalCount;
     }
     /**
      * 获取已检索总记录数
+     *
      * @param queryTotalCount
      */
     public void setQueryTotalCount(int queryTotalCount) {
         this.queryTotalCount = queryTotalCount;
     }
-    /**
-     * 初始化起始时间（毫秒）
-     */
+    /** 初始化起始时间（毫秒） */
     public void initStartTime() {
         this.startTime = System.currentTimeMillis();
     }
-    /**
-     * 初始化截止时间（毫秒）
-     */
+    /** 初始化截止时间（毫秒） */
     public void initEndTime() {
         this.endTime = System.currentTimeMillis();
     }
     /**
      * 获取毫秒格式的耗时信息
+     *
      * @return
      */
     public String getTimeIntervalByMilli() {
@@ -836,16 +845,15 @@ class HBasePageModel implements Serializable {
     }
     /**
      * 获取秒格式的耗时信息
+     *
      * @return
      */
     public String getTimeIntervalBySecond() {
-        double interval = (this.endTime - this.startTime)/1000.0;
+        double interval = (this.endTime - this.startTime) / 1000.0;
         DecimalFormat df = new DecimalFormat("#.##");
         return df.format(interval) + "秒";
     }
-    /**
-     * 打印时间信息
-     */
+    /** 打印时间信息 */
     public void printTimeInfo() {
         logger.info("起始时间：" + this.startTime);
         logger.info("截止时间：" + this.endTime);
@@ -853,6 +861,7 @@ class HBasePageModel implements Serializable {
     }
     /**
      * 获取HBase检索结果集合
+     *
      * @return
      */
     public List<Result> getResultList() {
@@ -860,6 +869,7 @@ class HBasePageModel implements Serializable {
     }
     /**
      * 设置HBase检索结果集合
+     *
      * @param resultList
      */
     public void setResultList(List<Result> resultList) {
@@ -867,18 +877,16 @@ class HBasePageModel implements Serializable {
     }
 }
 
-/**
- * 扩展 Put 类
- *  
- */
-class SocPut extends Put{
+/** 扩展 Put 类 */
+class SocPut extends Put {
 
     private Map<String, byte[]> map;
     private List<Map<String, byte[]>> data = new LinkedList<>();
 
     /**
      * 初始化方法
-     * @param row  rowKey 名称
+     *
+     * @param row rowKey 名称
      */
     public SocPut(byte[] row) {
         super(row);
@@ -898,42 +906,28 @@ class SocPut extends Put{
         return data;
     }
 
-    public int getDataSize(){
+    public int getDataSize() {
         return data.size();
     }
-
 }
 
 enum RandCodeEnum {
-    /**
-     * 混合字符串
-     */
-    ALL_CHAR("0123456789abcdefghijkmnpqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ"), // 去除小写的l和o这个两个不容易区分的字符；
-    /**
-     * 字符
-     */
+    /** 混合字符串 */
+    ALL_CHAR(
+            "0123456789abcdefghijkmnpqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ"), // 去除小写的l和o这个两个不容易区分的字符；
+    /** 字符 */
     LETTER_CHAR("abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ"),
-    /**
-     * 小写字母
-     */
+    /** 小写字母 */
     LOWER_CHAR("abcdefghijklmnopqrstuvwxyz"),
-    /**
-     * 数字
-     */
+    /** 数字 */
     NUMBER_CHAR("0123456789"),
-    /**
-     * 大写字符
-     */
+    /** 大写字符 */
     UPPER_CHAR("ABCDEFGHIJKLMNOPQRSTUVWXYZ"),
 
-    /**
-     * Hbase 离散前缀
-     */
+    /** Hbase 离散前缀 */
     HBASE_CHAR("123456789ABCDEF");
 
-    /**
-     * 待生成的字符串
-     */
+    /** 待生成的字符串 */
     private String charStr;
 
     private RandCodeEnum(final String charStr) {
@@ -956,7 +950,6 @@ enum RandCodeEnum {
         return charStr;
     }
 
-
     public String[] getHbaseKeys(int pNum, int b, boolean only) {
         Set<String> ts = new TreeSet<String>();
         int tss = 0;
@@ -973,7 +966,7 @@ enum RandCodeEnum {
     }
 
     public static void main(String[] args) {
-        String[] hbaseKeys = RandCodeEnum.HBASE_CHAR.getHbaseKeys(240,2,false);
+        String[] hbaseKeys = RandCodeEnum.HBASE_CHAR.getHbaseKeys(240, 2, false);
         for (String s : hbaseKeys) {
             System.out.println(s);
         }
